+++ packages/core/src/serializer.ts
@@ -3,7 +3,7 @@ import { asArray, keepDefineds, lazyMapGet, Nested3Map, uniquesAmong } from "@li
 import { asIds, metaPointerFor } from "./functions.js"
 import { Reader } from "./reading.js"
 import { Node } from "./types.js"
-import { builtinPropertyValueSerializer } from "./m3/builtins.js"
+import { lioncoreBuiltinsFacade } from "./m3/builtins.js"
 import { inheritsDirectlyFrom } from "./m3/functions.js"
 import {
     Classifier,
@@ -101,7 +101,7 @@ export const nodeSerializer = <NT extends Node>(reader: Reader<NT>, serializatio
 export const serializerWith = <NT extends Node>(configuration: SerializerConfiguration<NT>): Serializer<NT> => {
     const { reader } = configuration
     const propertyValueSerializer =
-        configuration.propertyValueSerializer ?? configuration.primitiveTypeSerializer ?? builtinPropertyValueSerializer
+        configuration.propertyValueSerializer ?? configuration.primitiveTypeSerializer ?? lioncoreBuiltinsFacade.propertyValueSerializer
     const serializeEmptyFeatures = configuration.serializeEmptyFeatures ?? true
 
     const languageKey2version2classifierKey2allFeatures: Nested3Map<Feature[]> = {}