+++ packages/test/src/core/serializer.test.ts
@@ -1,7 +1,5 @@
 import {
     Annotation,
-    builtinClassifiers,
-    builtinPrimitives,
     Concept,
     DynamicNode,
     dynamicReader,
@@ -9,6 +7,7 @@ import {
     EnumerationLiteral,
     Language,
     LanguageFactory,
+    lioncoreBuiltinsFacade,
     newPropertyValueSerializerRegistry,
     propertyValueSerializerFrom,
     Reference,
@@ -34,7 +33,7 @@ describe("serialization", () => {
     it("serializes node with custom primitive type, works when registering custom deserializer", () => {
         const builtinsPropertyValueSerializer = propertyValueSerializerFrom(
             newPropertyValueSerializerRegistry()
-                .set(builtinPrimitives.stringDataType, (value) => value as string)
+                .set(lioncoreBuiltinsFacade.primitiveTypes.stringDataType, (value) => value as string)
                 .set(dateDataType, (value) => {
                     const d = value as Date
                     return `${Number(d.getFullYear()).toString().padStart(4, "0")}-${Number(d.getMonth() + 1)
@@ -102,12 +101,14 @@ describe("serialization", () => {
             .to.eql(expectedSerializationChunk)
     })
 
+    const { inamed } = lioncoreBuiltinsFacade.classifiers
+
     it("serializes annotations", () => {
         const language = new Language("test language", "0", "test-language", "test-language")
         const annotatedConcept = new Concept(language, "Annotated", "Annotated", "Annotated", false)
-        annotatedConcept.implementing(builtinClassifiers.inamed)
+        annotatedConcept.implementing(inamed)
         const testAnnotation = new Annotation(language, "Annotation", "Annotation", "Annotation")
-        testAnnotation.implementing(builtinClassifiers.inamed)
+        testAnnotation.implementing(inamed)
         language.havingEntities(annotatedConcept, testAnnotation)
 
         const annotation = new TestNode("0", "Annotation")
@@ -230,13 +231,16 @@ describe("serialization", () => {
     })
 })
 
+
+const { primitiveTypes } = lioncoreBuiltinsFacade
+
 describe("serialization of empty (unset) values", () => {
     const factory = new LanguageFactory("serialization-language", "0", concatenator("-"), lastOf)
     const enumeration = factory.enumeration("enumeration")
     const concept = factory.concept("concept", false)
-    factory.property(concept, "stringProperty").ofType(builtinPrimitives.stringDataType).isOptional()
-    factory.property(concept, "integerProperty").ofType(builtinPrimitives.integerDataType).isOptional()
-    factory.property(concept, "booleanProperty").ofType(builtinPrimitives.booleanDataType).isOptional()
+    factory.property(concept, "stringProperty").ofType(primitiveTypes.stringDataType).isOptional()
+    factory.property(concept, "integerProperty").ofType(primitiveTypes.integerDataType).isOptional()
+    factory.property(concept, "booleanProperty").ofType(primitiveTypes.booleanDataType).isOptional()
     factory.property(concept, "enumProperty").ofType(enumeration).isOptional()
     factory.containment(concept, "containment").ofType(concept).isOptional()
     factory.containment(concept, "containments").ofType(concept).isOptional().isMultiple()