+++ packages/test/src/languages/library.ts
@@ -1,11 +1,11 @@
-import { builtinPrimitives, LanguageFactory } from "@lionweb/core"
+import { LanguageFactory, lioncoreBuiltinsFacade } from "@lionweb/core"
 import { chain, concatenator, lastOf } from "@lionweb/ts-utils"
 import { hasher } from "@lionweb/utilities"
 
 const factory = new LanguageFactory("library", "1", chain(concatenator("-"), hasher()), lastOf)
 export const libraryLanguage = factory.language
 
-const { integerDataType, stringDataType } = builtinPrimitives
+const { integerDataType, stringDataType } = lioncoreBuiltinsFacade.primitiveTypes
 
 const book = factory.concept("Book", false)
 const bookType = factory.enumeration("BookType")
@@ -31,3 +31,4 @@ factory.property(writer, "name").ofType(stringDataType).havingKey("library_Write
 factory.property(guideBookWriter, "countries").ofType(stringDataType)
 
 factory.property(specialistBookWriter, "subject").ofType(stringDataType)
+